+++ vpx_dsp/x86/highbd_idct4x4_add_sse4.c
@@ -18,11 +18,11 @@
 
 static INLINE void highbd_idct4(__m128i *const io) {
   const __m128i cospi_p16_p16 =
-      _mm_setr_epi32(cospi_16_64 << 2, 0, cospi_16_64 << 2, 0);
+      _mm_setr_epi32((int)cospi_16_64 << 2, 0, (int)cospi_16_64 << 2, 0);
   const __m128i cospi_p08_p08 =
-      _mm_setr_epi32(cospi_8_64 << 2, 0, cospi_8_64 << 2, 0);
+      _mm_setr_epi32((int)cospi_8_64 << 2, 0, (int)cospi_8_64 << 2, 0);
   const __m128i cospi_p24_p24 =
-      _mm_setr_epi32(cospi_24_64 << 2, 0, cospi_24_64 << 2, 0);
+      _mm_setr_epi32((int)cospi_24_64 << 2, 0, (int)cospi_24_64 << 2, 0);
   __m128i temp1[4], step[4];
 
   transpose_32bit_4x4(io, io);
